+++ src/pages/cv-common.ts
@@ -1,0 +1,19 @@
+export enum Command {
+  HELP = 'help',
+  WORK = 'work',
+  EDUCATION = 'education',
+  SKILLS = 'skills',
+  LANGUAGES = 'languages',
+  BASIC = 'basic',
+  AWARDS = 'awards',
+  PUBLICATIONS = 'publications',
+  INTERESTS = 'interests',
+  REFERENCES = 'references'
+}
+export const Commands: string[] = [
+  Command.HELP, Command.WORK,
+  Command.SKILLS, Command.LANGUAGES,
+  Command.BASIC, Command.EDUCATION,
+  Command.AWARDS, Command.PUBLICATIONS,
+  Command.INTERESTS, Command.REFERENCES
+];
