+++ src/pages/cv.css
@@ -7,4 +7,5 @@
     line-height: 1.46153846;
     text-align: left;
     -webkit-text-size-adjust: 100%;
+    width: 100%;
 }
