+++ src/pages/cv.tsx
@@ -2,6 +2,10 @@ import * as classnames from 'classnames';
 import * as React from 'react';
 
 import {
+  Command,
+  Commands
+} from './cv-common';
+import {
   CVAwards,
   CVBasics,
   CVHelp,
@@ -28,26 +32,6 @@ interface Model {
   raw: string;
   error: boolean;
 }
-
-export enum Command {
-  HELP = 'help',
-  WORK = 'work',
-  EDUCATION = 'education',
-  SKILLS = 'skills',
-  LANGUAGES = 'languages',
-  BASIC = 'basic',
-  AWARDS = 'awards',
-  PUBLICATIONS = 'publications',
-  INTERESTS = 'interests',
-  REFERENCES = 'references'
-}
-export const Commands: string[] = [
-  Command.HELP, Command.WORK,
-  Command.SKILLS, Command.LANGUAGES,
-  Command.BASIC, Command.EDUCATION,
-  Command.AWARDS, Command.PUBLICATIONS,
-  Command.INTERESTS, Command.REFERENCES
-];
 
 export class CVPage extends React.Component<Props, State> {
